entrypoint: Preserve ENV variables
A few ENV variables are set in the Docker file, most notably those
related to ccache.  The move to a docker entrypoint script  called
sudo to switch to the AOSP user and reset the ENV varibles.

Preserve these variables so that things set in the Dockerfile are
inherited by the environment as expected.